% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-extraction.R
\name{add_rownames_column}
\alias{add_rownames_column}
\title{Add row names as a column}
\usage{
add_rownames_column(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A data.frame with row names as a column
}
\description{
Adds row names as a column named "RowNxyz"
}
\author{
Kai Guo
}
\keyword{internal}
