% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.Venn}
\alias{as.data.frame.Venn}
\title{Convert a Venn object to a data frame}
\usage{
\method{as.data.frame}{Venn}(x, ...)
}
\arguments{
\item{x}{A Venn object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame with subset information
}
\description{
Converts a Venn object to a data frame for easier manipulation
}
\examples{
# Create a Venn object
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
res <- venndetail(list(A = A, B = B))

# Convert to data frame
df <- as.data.frame(res)
head(df)
}
\author{
Kai Guo
}
