% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{detail}
\alias{detail}
\title{detail generic}
\usage{
detail(x)
}
\arguments{
\item{x}{venn object}
}
\value{
vector include numbers belong to different sets
}
\description{
detail generic
}
\author{
Kai Guo
}
