% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dim.Venn}
\alias{dim.Venn}
\title{Get dimensions of a Venn object}
\usage{
\method{dim}{Venn}(x)
}
\arguments{
\item{x}{Venn object}
}
\value{
Integer vector of length 2 (rows, columns)
}
\description{
Returns the dimensions of the result slot in a Venn object.
}
