% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.add_colnames}
\alias{.add_colnames}
\title{Give first colname as RowNxyz}
\usage{
.add_colnames(x)
}
\arguments{
\item{x}{data frame}
}
\value{
return data frame with the first colnames change to "RowNxyz"
}
\description{
Give first colname as RowNxyz
}
