% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn-diagram.R
\name{.findOverlap}
\alias{.findOverlap}
\title{Find overlap between sets in a list}
\usage{
.findOverlap(setlist, xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{setlist}{list of character vectors.}

\item{xlim}{vector with 2 numbers, x axis limits for the venn diagram.}

\item{ylim}{vector with 2 numbers, y axis limits for the venn diagram.}
}
\value{
list with 2 items: a data.frame with information about the groups (sizes, coordinates, etc.),
        and a data.frame containing the x and y coordinates for the venn diagram
}
\description{
Find overlap between sets in a list
}
\keyword{internal}
