% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.make.table}
\alias{.make.table}
\title{make table for venndetail
modified from make.truth.table (VennDiagram)}
\usage{
.make.table(x)
}
\arguments{
\item{x}{A list with input groups}
}
\value{
A data frame with logical vector columns and 2 ^ length(x)-1 rows.
}
\description{
make table for venndetail
modified from make.truth.table (VennDiagram)
}
\author{
Kai Guo
}
