% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplot.R
\docType{method}
\name{get}
\alias{get}
\alias{get}
\title{get detail}
\usage{
get(object, ...)

\S4method{get}{venn}(object, group, ...)
}
\arguments{
\item{object}{venn object}

\item{...}{additional parameters}

\item{group}{group you want extract}

\item{x}{venn object}
}
\value{
return dataframe and print the header of dataframe
}
\description{
get generic
}
\examples{
\dontrun{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
get(res,"A")
}
}
\author{
Kai Guo
}
