% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.set}
\alias{get.set}
\title{get set}
\usage{
get.set(x, tab, sep = "_")
}
\arguments{
\item{x}{A list with input groups}

\item{tab}{TRUE or FALSE table}

\item{sep}{Character string used to separate the terms when concatenating
group names into new column names (colnames).}
}
\value{
A list with groups and details
}
\description{
get set
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
x = list(A = A, B = B, C = C)
tab <- make.table(x = list(A = A, B = B, C = C))
out <- get.set(x = x, tab = tab)
}
\author{
Kai Guo
}
