% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getFeature}
\alias{getFeature}
\title{get feature based on venn results}
\usage{
getFeature(object, ...)
}
\arguments{
\item{object}{venn object}

\item{...}{additional parameters}
}
\value{
dataframe with all venn information and details from input
}
\description{
getFeature generic
}
\author{
Kai Guo
}
