% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{head.Venn}
\alias{head.Venn}
\title{Print the head of Venn object}
\usage{
\method{head}{Venn}(x, n = 6L, ...)
}
\arguments{
\item{x}{Venn object}

\item{n}{number of rows to display}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Print the head of Venn object
}
