% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadVenn}
\alias{loadVenn}
\title{Load a Venn object from a file}
\usage{
loadVenn(file)
}
\arguments{
\item{file}{File name to load from}
}
\value{
A Venn object
}
\description{
Loads a Venn object from an RDS file
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
res <- venndetail(list(A = A, B = B))
tmp <- tempfile(fileext = ".rds")
saveVenn(res, tmp)
res2 <- loadVenn(tmp)
}
\author{
Kai Guo
}
