% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make.table}
\alias{make.table}
\title{make table for venndetail
modified from make.truth.table (VennDaigram)}
\usage{
make.table(x)
}
\arguments{
\item{x}{A list with input groups}
}
\value{
A data frame with logical vector columns and 2 ^ length(x) rows.
}
\description{
make table for venndetail
modified from make.truth.table (VennDaigram)
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
tab <- make.table(x = list(A = A, B = B, C = C))
}
\author{
Kai Guo
}
