% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndetail.R
\name{make_truth_table}
\alias{make_truth_table}
\title{Create a truth table for set combinations}
\usage{
make_truth_table(x)
}
\arguments{
\item{x}{A list of vectors}
}
\value{
A data frame with logical columns for each set and rows for each combination
}
\description{
Creates a logical matrix representing all possible combinations of sets
}
\author{
Kai Guo
}
\keyword{internal}
