% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uset-plot.R
\name{makeContent.upset_grob}
\alias{makeContent.upset_grob}
\title{Draw method for upset_grob objects}
\usage{
\method{makeContent}{upset_grob}(x)
}
\arguments{
\item{x}{An upset_grob object}

\item{recording}{Logical for recording}
}
\value{
Draws the upset plot
}
\description{
Internal method to draw upset_grob using grid viewports
}
\keyword{internal}
