% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{merge}
\alias{merge}
\title{merge two or more venn object}
\usage{
merge(object, ...)
}
\arguments{
\item{object}{list of venn object}
}
\value{
venn object
}
\description{
Merge two or more venn object by group name
}
\author{
Kai Guo
}
