% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,list-method}
\title{Merge two or more venn object by group name}
\usage{
merge(object, ignore.case = FALSE, useupper = TRUE, plot = FALSE,
  ...)

\S4method{merge}{list}(object, ignore.case = FALSE, useupper = TRUE,
  plot = FALSE, ...)
}
\arguments{
\item{object}{list of venn objects}

\item{ignore.case}{ignore case of group name (default: FALSE)}

\item{useupper}{use uppercase for all group name (default: TRUE)}

\item{plot}{plot figure or not (default: FALSE)}
}
\value{
venn object
}
\description{
Merge will combine multiple venn diagrams to allow comparison between multiple groups
}
\author{
Kai Guo
}
