% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{names.Venn}
\alias{names.Venn}
\title{Extract subset names from a Venn object}
\usage{
\method{names}{Venn}(x)
}
\arguments{
\item{x}{A Venn object}
}
\value{
A character vector of subset names
}
\description{
Returns the names of all subsets in a Venn object
}
\examples{
# Create a Venn object
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
res <- venndetail(list(A = A, B = B))

# Get subset names
names(res)
}
\author{
Kai Guo
}
