% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uset-plot.R
\name{print.upset_grob}
\alias{print.upset_grob}
\title{Print method for upset_grob objects}
\usage{
\method{print}{upset_grob}(x, ...)
}
\arguments{
\item{x}{An upset_grob object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for upset_grob objects
}
