% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{rowjoin,data.frame,data.frame-method}
\alias{rowjoin,data.frame,data.frame-method}
\title{join by rownames}
\usage{
\S4method{rowjoin}{data.frame,data.frame}(x, y, fun = "full_join")
}
\arguments{
\item{x}{data.frame x}

\item{y}{data.frame y}

\item{fun}{join type}
}
