% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-extraction.R
\name{set_rownames}
\alias{set_rownames}
\title{Set row names from a column}
\usage{
set_rownames(x, col)
}
\arguments{
\item{x}{A data.frame}

\item{col}{Column index or name to use for row names}
}
\value{
A data.frame with row names from the specified column
}
\description{
Sets row names from a specified column
}
\author{
Kai Guo
}
\keyword{internal}
