% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/plot.R, R/utils.R
\name{setcolor}
\alias{setcolor}
\title{return colors with given a vector}
\usage{
setcolor(n, palette = "default")

setcolor(n, palette = "default")

setcolor(n, palette = "default")
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Type of palette: "default", "categorical", "sequential", or "diverging"}

\item{x}{Number of color}
}
\value{
color vector

A vector of colors

A vector of colors
}
\description{
Setcolor will provide a list of color vectors based on
the number used as an input.

Returns a color palette for Venn diagrams

Returns a color palette for Venn diagrams
}
\examples{
mycol <- setcolor(10)
mycol
# Get 5 colors from the default palette
cols <- setcolor(5)

# Get 10 colors from the categorical palette
cols <- setcolor(10, palette = "categorical")
# Get 5 colors from the default palette
cols <- setcolor(5)

# Get 10 colors from the categorical palette
cols <- setcolor(10, palette = "categorical")
}
\author{
Kai Guo
}
