% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/summary.R
\name{summary.Venn}
\alias{summary.Venn}
\title{Return summary information about a Venn object}
\usage{
\method{summary}{Venn}(object, ...)

\method{summary}{Venn}(object, ...)
}
\arguments{
\item{object}{A Venn object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the Venn object

Invisibly returns the Venn object
}
\description{
Summary method for Venn objects

Summary method for Venn objects
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
summary(res)
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
summary(res)
}
\author{
Kai Guo
}
