% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDetail-package.R
\docType{data}
\name{T2DM}
\alias{T2DM}
\title{T2DM Dataset}
\format{
A list of data frames with five columns each
}
\usage{
T2DM
}
\description{
T2DM data are differential expression genes (DEGs) with
  annotation from the publication by Hinder et al.
}
\keyword{datasets}
