% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndetail.R
\name{venndetail}
\alias{venndetail}
\title{Extract shared and unique sets}
\usage{
venndetail(x, plot = TRUE, filename = NULL, type = "venn",
  col = "black", sep = "_", mycol = c("dodgerblue", "goldenrod1",
  "darkorange1", "seagreen3", "orchid3"), cat.cex = 1.5, alpha = 0.5,
  cex = 2, cat.fontface = "bold", margin = 0.05, abbr = FALSE,
  minlength = 3, abbr.method = "both.sides", text.scale = c(1.5, 1.5,
  1.5, 1.5, 1.5, 1.5))
}
\arguments{
\item{x}{A list of variables with group names.}

\item{plot}{Boolean indicating whether to draw a plot  (default:TRUE).}

\item{filename}{Filename for output figure.}

\item{type}{One of c(‘venn’, ‘vennpie’, ‘upset’) to specify the type of
plot (default: venn).}

\item{col}{Character vector giving the color of the circles.}

\item{sep}{The separation character used for column names of detailed
output files (default: ‘_’).}

\item{mycol}{Character vector giving the filled color for circles.}

\item{cat.cex}{Numeric vector giving the size of the category names.}

\item{alpha}{A number giving the transparency value.}

\item{cex}{Vector giving the size for each area label}

\item{cat.fontface}{A character giving the fontface (font style) for
category name.}

\item{margin}{Number giving the amount of whitespace around the diagram in
grid units}

\item{abbr}{Boolean indicating whether to abbreviate set names
(default: FALSE).}

\item{minlength}{minmal length for the set name.}

\item{abbr.method}{abbreviation method (default: both side).}

\item{text.scale}{Numeric vector of text sizes for upset diagram
(ylab,yaxis,xlab,set name,xaxis,insection).}
}
\value{
venn object and figures
}
\description{
Extracts shared and unique elements from groups provided to
the function.This base function will create a formal class venn object and
can also graphically plot the amount of objects in each group. The plot will
be in the form of a traditional venn diagram as default.And users can also
use vennpie or upset methods to display the result.
}
\details{
Extract shared and unique sets
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
}
\author{
Kai Guo
}
