% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/vennpie.R
\docType{methods}
\name{vennpie}
\alias{vennpie}
\alias{vennpie}
\title{pie plot show shared and unique part}
\usage{
vennpie(object, ...)

\S4method{vennpie}{venn}(object, group = NULL, color = NULL,
  revcolor = "lightgrey", any = NULL, show.number = TRUE, show.x = TRUE,
  sep = "_", log = FALSE, base = NULL, percentage = FALSE, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{set name you want display}

\item{color}{vector of color you want use for the group}

\item{revcolor}{backgroup color}

\item{show.number}{display the element number of the group or not}

\item{sep}{separate for new colnames}

\item{log}{use log transform or not}

\item{base}{log base}

\item{percentage}{display percentage format or not}
}
\value{
plot
}
\description{
vennpie generic
}
\examples{
\dontrun{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
dA=data.frame(A=A,"FC"=rnorm(40))
dB=data.frame(B=B,"FC"=rnorm(60))
dC=data.frame(C=C,"FC"=rnorm(40))
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
vennpie(res)
}
}
\author{
Kai Guo
}
