% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{$.Venn}
\alias{$.Venn}
\title{Extract column from a Venn object}
\usage{
\method{$}{Venn}(x, name)
}
\arguments{
\item{x}{Venn object}

\item{name}{Column name to extract}
}
\value{
Vector containing the specified column
}
\description{
Extract a column from the result slot of a Venn object using $ notation.
}
