% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn-pie.R
\name{create_interactive_vennpie}
\alias{create_interactive_vennpie}
\title{Create an interactive vennpie chart with plotly}
\usage{
create_interactive_vennpie(
  object,
  subset = NULL,
  any = NULL,
  color = NULL,
  revcolor = "lightgrey",
  title = NULL
)
}
\arguments{
\item{object}{A Venn object}

\item{subset}{Character vector of subset names to highlight}

\item{any}{Highlight subsets shared by exactly this many sets}

\item{color}{Optional vector of colors for the subsets}

\item{revcolor}{Color for non-highlighted subsets}

\item{title}{Optional plot title}
}
\value{
A plotly object
}
\description{
Creates a simple pie chart visualization for interactive exploration of set intersections
}
\author{
Kai Guo
}
