% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\name{detail}
\alias{detail}
\alias{detail.venn}
\title{detail generic}
\usage{
detail(x)

\method{detail}{venn}(x)
}
\arguments{
\item{x}{venn object}
}
\value{
vector include numbers belong to different sets
}
\description{
detail generic

detail venn object
}
\examples{
{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=FALSE)
detail(res)
}
}
\author{
Kai Guo
}
