% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\name{detail}
\alias{detail}
\alias{detail.venn}
\title{Detail function provides a way to display the amount of members in
each group}
\usage{
detail(object)

\method{detail}{venn}(object)
}
\arguments{
\item{object}{venn object}
}
\value{
Numeric vector with group names and amounts for each group
}
\description{
The objective of this function is to summarizes the overlaps
across groups identified by venndetail without creating diagram.
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
detail(res)
}
\author{
Kai Guo
}
