% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\name{detail}
\alias{detail}
\alias{detail.venn}
\title{The detail function provides a way to display the amount of members in each group}
\usage{
detail(x)

\method{detail}{venn}(x)
}
\arguments{
\item{x}{venn object}
}
\value{
Vector including group names and total member amounts for each group
}
\description{
The objective of this function is to allow insight into the amount of overlap between groups identified by
venndetail without the need to print a diagram. The groups displayed will include all comparisons such as
shared, pairwise, and unique.
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=FALSE)
detail(res)
}
\author{
Kai Guo
}
