% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{detail}
\alias{detail}
\alias{detail,Venn-method}
\title{Detail function provides a way to display the amount of members in
each group}
\usage{
detail(object)

\S4method{detail}{Venn}(object)
}
\arguments{
\item{object}{Venn object}
}
\value{
Numeric vector with set names and amounts for each set
}
\description{
The objective of this function is to summarizes the overlaps
across groups identified by venndetail without creating diagram.
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
detail(res)
}
\author{
Kai Guo
}
