% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.Venn}
\alias{dim.Venn}
\title{Get the dimensions of a Venn object}
\usage{
\method{dim}{Venn}(x)
}
\arguments{
\item{x}{A Venn object}
}
\value{
A numeric vector with the number of rows and columns
}
\description{
Returns the dimensions of the result data frame
}
\examples{
# Create a Venn object
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
res <- venndetail(list(A = A, B = B))

# Get dimensions
dim(res)
}
\author{
Kai Guo
}
