% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplot.R
\docType{methods}
\name{dplot}
\alias{dplot}
\alias{dplot}
\title{detail plot}
\usage{
dplot(object, order = FALSE, ...)

\S4method{dplot}{venn}(object, order = FALSE, textsize = 5, ...)
}
\arguments{
\item{object}{venn object}
}
\value{
plot
}
\description{
dplot generic
}
\examples{
\dontrun{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
dplot(res,order=T,textsize=3)
}
}
\author{
Kai Guo
}
