% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplot.R
\docType{methods}
\name{Get}
\alias{Get}
\alias{Get venn}
\title{Get the contents of a specific set from venndetail object}
\usage{
Get(object, group, ...)

\S4method{Get}{venn}(object, group, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{Character vector giving the set names}
}
\value{
Specific set information
}
\description{
Get the contents of a specific set from venndetail object
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
Get(res,"A")
}
\author{
Kai Guo

Kai Guo
}
