% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplot.R
\docType{methods}
\name{get}
\alias{get}
\title{Get the contents of a specified group from venndetail object}
\usage{
get(object, group, ...)

\S4method{get}{venn}(object, group, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{user defined specified group}
}
\value{
return dataframe and print the header of dataframe
}
\description{
Get the contents of a specified group from venndetail object
}
\examples{
A<-sample(1:100,40,replace = FALSE);
B<-sample(1:100,60,replace = FALSE);
C<-sample(1:100,40,replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
get(res,"A")
}
\author{
Kai Guo

Kai Guo
}
