% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot.R
\name{get}
\alias{get}
\alias{get.venn}
\title{Get the contents of a specific set from venndetail object}
\usage{
get.venn(object, group, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{Character vector giving the set names}
}
\value{
Specific set information
}
\description{
Get the contents of a specific set from venndetail object
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
get(res,"A")
}
\author{
Kai Guo
}
