% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getFeature.R
\docType{methods}
\name{getFeature}
\alias{getFeature}
\alias{getFeature}
\title{get feature based on venn results}
\usage{
getFeature(object, group, rlist, userowname = TRUE, gind = NULL,
  sep = "_", ...)

\S4method{getFeature}{venn}(object, group, rlist, userowname = TRUE,
  gind = NULL, sep = "_", ...)
}
\arguments{
\item{object}{venn object}

\item{group}{group you want used}

\item{rlist}{list of detail dataframe with all information}

\item{userowname}{use rowname to join dataframe or not}

\item{gind}{id name you want use to extract for each data.frame (set if userowname=F)}

\item{sep}{separate for new colnames}
}
\value{
dataframe with all venn information and details from input
}
\description{
getFeature generic
}
\examples{
\dontrun{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
dA=data.frame(A=A,"FC"=rnorm(40))
dB=data.frame(B=B,"FC"=rnorm(60))
dC=data.frame(C=C,"FC"=rnorm(40))
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
getFeature(res,group="Shared",rlist=list(dA,dB,dC),userowname=F)
}
}
\author{
Kai Guo
}
