% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getFeature.R
\docType{methods}
\name{getFeature}
\alias{getFeature}
\title{getFeature provide a way to combine list of user supplied data.frame with venndetail results}
\usage{
getFeature(object, group, rlist, userowname = TRUE, gind = NULL,
  sep = "_", ...)

\S4method{getFeature}{venn}(object, group, rlist, userowname = TRUE,
  gind = NULL, sep = "_", ...)
}
\arguments{
\item{object}{venn object}

\item{group}{user defined group want to extract}

\item{rlist}{list of user supplied dataframe to combine with venndetail result}

\item{userowname}{use rowname to join dataframe or not (default: TRUE)}

\item{gind}{column name or index for each data.frame user want to use (userowname=FALSE)}

\item{sep}{separate of the new colnames for the return data.frame}
}
\value{
data.frame with subset information and detail from user supplied data.frame
}
\description{
Getfeature allows users a method of extracting members of groups from venndetail in table format
that can also include accompanying information from a data frame provided in the rlist argument
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
dA=data.frame(A=A,"FC"=rnorm(40))
dB=data.frame(B=B,"FC"=rnorm(60))
dC=data.frame(C=C,"FC"=rnorm(40))
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
rhs<-getFeature(res,group="Shared",rlist=list(dA,dB,dC),userowname=FALSE,gind=rep(1,3))
}
\author{
Kai Guo
}
