% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{head.Venn}
\alias{head.Venn}
\alias{tail.Venn}
\title{Return the first or last parts of a Venn object}
\usage{
\method{head}{Venn}(x, n = 6L, ...)

\method{tail}{Venn}(x, n = 6L, ...)
}
\arguments{
\item{x}{Venn object}

\item{n}{number of rows to display}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\value{
A data.frame with the first n rows

A data.frame with the last n rows
}
\description{
Methods to extract the first or last parts of a Venn object's result slot.
}
