% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadVenn}
\alias{loadVenn}
\title{Load a Venn object from a file}
\usage{
loadVenn(file)
}
\arguments{
\item{file}{File name to load from}
}
\value{
A Venn object
}
\description{
Loads a Venn object from an RDS file
}
\examples{
\dontrun{
# Load a saved Venn object
res <- loadVenn("my_venn.rds")

# Plot the loaded object
plot(res)
}
}
\author{
Kai Guo
}
