% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{merge,list-method}
\alias{merge,list-method}
\title{merge two or more venn object by group name}
\usage{
\S4method{merge}{list}(object, ingore.case = F, useupper = T)
}
\arguments{
\item{object}{list of venn object}

\item{ingore.case}{ingore case of group name}

\item{useupper}{use uppercase for all group name}
}
\description{
merge two or more venn object by group name
}
\author{
Kai Guo
}
