% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{merge}
\alias{merge}
\title{merge two or more venn object}
\usage{
merge(object, ingore.case = F, useupper = T, plot = FALSE, ...)
}
\arguments{
\item{object}{list of venn object}

\item{ingore.case}{ingore case of group name}

\item{useupper}{use uppercase for all group name}

\item{plot}{plot figure or not}
}
\value{
venn object
}
\description{
Merge two or more venn object by group name
}
\author{
Kai Guo
}
