% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,list-method}
\title{merge two or more venn object}
\usage{
merge(object, ingore.case = FALSE, useupper = TRUE, plot = FALSE,
  ...)

\S4method{merge}{list}(object, ignore.case = FALSE, useupper = TRUE,
  plot = FALSE, ...)
}
\arguments{
\item{object}{list of venn object}

\item{ingore.case}{ingore case of group name}

\item{useupper}{use uppercase for all group name}

\item{plot}{plot figure or not}

\item{ignore.case}{ignore case of group name}

\item{object}{list of venn object}

\item{useupper}{use uppercase for all group name}

\item{plot}{plot figure or not}
}
\value{
venn object
}
\description{
Merge two or more venn object by group name

Merge will combine multiple venn diagrams to allow comparison between multiple groups
}
\author{
Kai Guo
}
