% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.venn}
\alias{plot.venn}
\title{Plot venn object}
\usage{
\method{plot}{venn}(x, type = "venn", col = "black", sep = "_",
  mycol = c("dodgerblue", "goldenrod1", "darkorange1", "seagreen3",
  "orchid3"), cat.cex = 1.5, alpha = 0.5, cex = 2,
  cat.fontface = "bold", margin = 0.05, text.scale = c(1.5, 1.5, 1.5,
  1.5, 1.5, 1.5), filename = NULL, piecolor = NULL,
  revcolor = "lightgrey", any = NULL, show.number = TRUE,
  show.x = TRUE, log = FALSE, base = NULL, percentage = FALSE,
  sets.x.label = "Set Size", mainbar.y.label = "Intersection Size",
  nintersects = 40, ...)
}
\arguments{
\item{type}{use venn,vennpie or upset (default: venn)}

\item{col}{color for the circle}

\item{sep}{separation character used for the group and numbers}

\item{mycol}{color for VennDiagram graphic}

\item{cat.cex}{font size}

\item{alpha}{transparency value}

\item{cat.fontface}{font type}

\item{text.scale}{vector of text size for upset (ylab,yaxis,xlab,group name,xaxis,insection)}

\item{filename}{output filename}

\item{sets.x.label}{The x-axis label of the set size bar plot}

\item{mainbar.y.label}{The y-axis label of the intersection size bar plot}

\item{nintersects}{Number of intersections to plot. If set to NA, all intersections will be plotted.}

\item{abbr}{abbreviate name or not (default: FALSE)}

\item{minlength}{set minmal length for names}

\item{abbr.method}{abbreviation method(default: both side)}
}
\value{
different type of graphics based on user chose
}
\description{
Plot allows users to graphically display the groups and overlap between groups
in their venn class object through a variety of graph types such as a bar plot, traditional venn,
or venn pie chart.
}
\examples{
A<-sample(1:100,40,replace = FALSE);
B<-sample(1:100,60,replace = FALSE);
C<-sample(1:100,40,replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=FALSE)
plot(res,type="venn")
}
\author{
Kai Guo
}
