% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.Venn}
\alias{plot.Venn}
\title{Plot Venn object}
\usage{
\method{plot}{Venn}(x, type = "venn", col = "black", sep = "_",
  mycol = c("dodgerblue", "goldenrod1", "darkorange1", "seagreen3",
  "orchid3"), cat.cex = 1.5, alpha = 0.5, cex = 2,
  cat.fontface = "bold", margin = 0.05, text.scale = c(1.5, 1.5, 1.5,
  1.5, 1.5, 1.5), filename = NULL, piecolor = NULL,
  revcolor = "lightgrey", any = NULL, show.number = TRUE,
  show.x = TRUE, log = FALSE, base = NULL, percentage = FALSE,
  sets.x.label = "Set Size", mainbar.y.label = "Intersection Size",
  nintersects = 40, abbr = FALSE, abbr.method = "both.sides",
  minlength = 3, ...)
}
\arguments{
\item{x}{Venn object}

\item{type}{Use venn, vennpie or upset (default: venn)}

\item{col}{Character vector giving the color of the circles.}

\item{sep}{Character string used to separate the terms when concatenating
group names into new column names (colnames)(vennpie).}

\item{mycol}{Character vector giving the filled color for
VennDiagram circles.}

\item{cat.cex}{Numeric vector giving the size of the category names.}

\item{alpha}{A number giving the transparency value.}

\item{cex}{A numerical value giving the text size for venndiagram}

\item{cat.fontface}{A character giving the fontface (font style) for
category name.}

\item{margin}{Number giving the amount of whitespace around the diagram in
grid units}

\item{text.scale}{Numeric vector of text sizes for upset diagram
(ylab, yaxis, xlab, subset name, xaxis, insection).}

\item{filename}{Filename for output figure.}

\item{piecolor}{Character vector giving the colors of the subsets(vennpie).}

\item{revcolor}{Character giving the color for the non-selected
subsets(vennpie).}

\item{any}{Number to indicate selected subsets, such as 1 means any unique
subsets, 2 means any subsets shared by two groups(vennpie).}

\item{show.number}{Boolean indicating whether to display the
element numbers of the subsets or not (default: TRUE)(vennpie).}

\item{show.x}{Boolean indicating whether to show subset labels outside the
circle (default: TRUE)(vennpie).}

\item{log}{Boolean indicating whether to transform the data in
log scale(vennpie).}

\item{base}{Base value for log transformation(vennpie).}

\item{percentage}{Boolean indicating whether to display subset percentages
(default: FALSE)(vennpie).}

\item{sets.x.label}{x-axis label (upset)}

\item{mainbar.y.label}{y-axis label (upset)}

\item{nintersects}{Number of intersections to plot. If subset to NA, all
intersections will be plotted.}

\item{abbr}{Boolean indicating whether to abbreviate subset names
(default: FALSE).}

\item{abbr.method}{a character string specifying the method used.
Partial matches allowed. (default: both side).}

\item{minlength}{Minmal length for the subset name.}

\item{...}{further arguments passed to or from other methods}
}
\value{
different type of graphics based on user chose
}
\description{
The plot function allows users to graphically display the
groups and overlap between groups in their venn class object through a
variety of graph types such as a bar plot, traditional venn,
or venn pie chart.
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
plot(res, type = "venn")
}
\author{
Kai Guo
}
