% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{result}
\alias{result}
\title{result generic}
\usage{
result(x)
}
\arguments{
\item{x}{venn object}
}
\value{
return dataframe and print header of dataframe
}
\description{
result generic
}
\author{
Kai Guo
}
