% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\name{result}
\alias{result}
\alias{result.venn}
\title{Extract the result from venn object}
\usage{
result(object)

\method{result}{venn}(object)
}
\arguments{
\item{object}{venn object}
}
\value{
return dataframe and print header of dataframe
}
\description{
Result will return output in a table format including the
contents of the sets included in the venndetail object
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
result <- result(res)
}
\author{
Kai Guo
}
