% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\name{result}
\alias{result}
\alias{result.venn}
\title{result generic}
\usage{
result(x)

\method{result}{venn}(x)
}
\arguments{
\item{x}{venn object}
}
\value{
return dataframe and print header of dataframe
}
\description{
result generic

extract result from venn object
}
\examples{
{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=FALSE)
result<-result(res)
}
}
\author{
Kai Guo
}
