% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{result}
\alias{result}
\alias{result,Venn-method}
\title{Extract the result from venn object}
\usage{
result(object, wide = FALSE)

\S4method{result}{Venn}(object, wide = FALSE)
}
\arguments{
\item{object}{Venn object}

\item{wide}{Boolean indicating whether to return wide format(default:FALSE)}
}
\value{
return dataframe and print header of dataframe
}
\description{
Result will return output in a table format including the
contents of the subsets included in the venndetail object
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
result <- result(res)
}
\author{
Kai Guo
}
