% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{rowjoin}
\alias{rowjoin}
\alias{rowjoin,data.frame,data.frame-method}
\title{join data.frame based on rownames}
\usage{
rowjoin(x, y, ...)

\S4method{rowjoin}{data.frame,data.frame}(x, y, fun = "full_join")
}
\arguments{
\item{x}{dataframe x}

\item{y}{dataframe y}

\item{fun}{join type}

\item{x}{data.frame x}

\item{y}{data.frame y}

\item{fun}{join type}
}
\value{
dataframe with join results
}
\description{
rowjoin generic
}
\author{
Kai Guo

Kai Guo
}
