% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowjoin}
\alias{rowjoin}
\title{rowjoin generic}
\usage{
rowjoin(x, y, ...)
}
\arguments{
\item{x}{dataframe x}

\item{y}{dataframe y}

\item{fun}{join type}
}
\value{
dataframe with join results
}
\description{
rowjoin generic
}
\author{
Kai Guo
}
