% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/misc.R
\docType{methods}
\name{rowjoin}
\alias{rowjoin}
\alias{rowjoin,data.frame,data.frame-method}
\title{Join data.frame based on rownames}
\usage{
rowjoin(x, y, fun = "fun_join")

\S4method{rowjoin}{data.frame,data.frame}(x, y, fun = "full_join")
}
\arguments{
\item{x}{data.frame x}

\item{y}{data.frame y}

\item{fun}{Different join format: left_join, full_join, right_join
(default:full_join)}
}
\value{
dataframe with join results
}
\description{
join two dataframes by rownames
}
\examples{
library(dplyr)
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
dA <- data.frame(A = A, "FC" = rnorm(40))
dB <- data.frame(B = B, "FC" = rnorm(60))
rownames(dA) <- A
rownames(dB) <- B
rowjoin(dA, dB)
}
\author{
Kai Guo
}
