% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setcolor}
\alias{setcolor}
\title{set color with given a vector}
\usage{
setcolor(x)
}
\arguments{
\item{x}{length of color}
}
\value{
color palette
}
\description{
set color with given a vector
}
\examples{
{
mycol<-setcolor(10)
}
}
\author{
Kai Guo
}
