% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/utils.R
\name{setcolor}
\alias{setcolor}
\title{Set color palette for Venn diagrams}
\usage{
setcolor(n, palette = "default")

setcolor(n, palette = "default")
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Type of palette: "default", "categorical", "sequential", or "diverging"}
}
\value{
A vector of colors

A vector of colors
}
\description{
Returns a color palette for Venn diagrams

Returns a color palette for Venn diagrams
}
\examples{
# Get 5 colors from the default palette
cols <- setcolor(5)

# Get 10 colors from the categorical palette
cols <- setcolor(10, palette = "categorical")
# Get 5 colors from the default palette
cols <- setcolor(5)

# Get 10 colors from the categorical palette
cols <- setcolor(10, palette = "categorical")
}
\author{
Kai Guo
}
