% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setcolor}
\alias{setcolor}
\title{return colors with given a vector}
\usage{
setcolor(n, palette = "default")
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Type of palette: "default", "categorical", "sequential", or "diverging"}
}
\value{
color vector
}
\description{
Setcolor will provide a list of color vectors based on
the number used as an input.
}
\examples{
mycol <- setcolor(10)
mycol
}
\author{
Kai Guo
}
