% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setcolor}
\alias{setcolor}
\title{return colors with given a vector}
\usage{
setcolor(x)
}
\arguments{
\item{x}{Number of color}
}
\value{
color vector
}
\description{
Setcolor will provide a list of color vectors based on
the number used as an input.
}
\examples{
mycol <- setcolor(10)
mycol
}
\author{
Kai Guo
}
