% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/summary.R
\name{show,Venn-method}
\alias{show,Venn-method}
\title{Show method for Venn objects}
\usage{
\S4method{show}{Venn}(object)

\S4method{show}{Venn}(object)
}
\arguments{
\item{object}{A Venn object}
}
\value{
Invisibly returns the Venn object

Invisibly returns the Venn object
}
\description{
Display method for Venn objects

Display method for Venn objects
}
\author{
Kai Guo
}
