% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot.R
\docType{methods}
\name{show}
\alias{show}
\title{Show the summary of venn object}
\usage{
\S4method{show}{venn}(object)
}
\arguments{
\item{object}{venn object}
}
\value{
summary information for the venn object
}
\description{
This function provides a summary of the venn object, including
a full results and sets as well as an summary information.
}
\examples{
A <- sample(1:100, 40, replace = FALSE);
B <- sample(1:100, 60, replace = FALSE);
C <- sample(1:100, 40, replace = FALSE);
res <- venndetail(list(A = A,B = B,C = C), plot = TRUE)
show(res)
}
\author{
Kai Guo
}
