% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot.R
\docType{methods}
\name{show Venn}
\alias{show Venn}
\alias{show,Venn-method}
\title{Show the summary of venn object}
\usage{
\S4method{show}{Venn}(object)
}
\arguments{
\item{object}{venn object}
}
\value{
summary information for the venn object
}
\description{
This function provides a summary of the venn object, including
a full results and subsets as well as an summary information.
}
\examples{
A <- sample(1:100, 40, replace = FALSE)
B <- sample(1:100, 60, replace = FALSE)
C <- sample(1:100, 40, replace = FALSE)
res <- venndetail(list(A = A, B = B, C = C))
show(res)

}
\author{
Kai Guo
}
