% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{[.Venn}
\alias{[.Venn}
\title{Subset a Venn object}
\usage{
\method{[}{Venn}(x, i, j)
}
\arguments{
\item{x}{Venn object}

\item{i}{Row indices}

\item{j}{Column indices}
}
\value{
Subset of the result data.frame
}
\description{
Extract elements from the result slot of a Venn object using bracket notation.
}
