% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{venn-class}
\alias{venn-class}
\alias{plot,}
\alias{venn-method}
\alias{summary,}
\title{Class 'venn'
This class includes all information from venndetail}
\description{
Class 'venn'
This class includes all information from venndetail
}
\section{Slots}{

\describe{
\item{\code{input}}{orginal input datasets}

\item{\code{raw}}{summary of the input datasets}

\item{\code{sep}}{separation character}

\item{\code{GroupNames}}{input group names}

\item{\code{result}}{shared or unique sets}

\item{\code{detail}}{shared of unique number belongs to each sets}
}}

\author{
Kai Guo
}
\keyword{classes}
