% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{venn-class}
\alias{venn-class}
\alias{plot,}
\alias{venn-method}
\alias{summary,}
\title{Class 'venn'
This class include all information from venndetail}
\description{
Class 'venn'
This class include all information from venndetail
}
\section{Slots}{

\describe{
\item{\code{GroupNames}}{group names input}

\item{\code{result}}{intersect and difference for groups}

\item{\code{detail}}{number belongs to each group}
}}

\author{
Kai Guo
}
\keyword{classes}
