% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{Venn-class}
\alias{Venn-class}
\title{Class 'Venn'
This class includes all information from venndetail}
\description{
Class 'Venn'
This class includes all information from venndetail
}
\section{Slots}{

\describe{
\item{\code{input}}{orginal input datasets}

\item{\code{raw}}{summary of the input datasets}

\item{\code{sep}}{separation character}

\item{\code{GroupNames}}{input group names}

\item{\code{result}}{shared or unique sets}

\item{\code{detail}}{shared of unique number belongs to each sets}

\item{\code{wide}}{result in wide format}
}}

\author{
Kai Guo
}
\keyword{classes}
