% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndetail.R
\name{venndetail}
\alias{venndetail}
\title{Extract shared and unique sets}
\usage{
venndetail(x, plot = TRUE, filename = NULL, type = "venn",
  col = "black", sep = "_", mycol = c("dodgerblue", "goldenrod1",
  "darkorange1", "seagreen3", "orchid3"), cat.cex = 1.5, alpha = 0.5,
  cex = 2, cat.fontface = "bold", margin = 0.05, abbr = FALSE,
  minlength = 3, abbr.method = "both.sides", text.scale = c(1.5, 1.5,
  1.5, 1.5, 1.5, 1.5))
}
\arguments{
\item{x}{list of variables with group names}

\item{plot}{plot the graphic (default:TRUE)}

\item{filename}{output filename}

\item{type}{use venn,vennpie or upset (default: venn)}

\item{col}{color for the circle}

\item{sep}{separation character used for the group and numbers}

\item{mycol}{color for VennDiagram graphic}

\item{cat.cex}{font size}

\item{alpha}{transparency value}

\item{cat.fontface}{font type}

\item{abbr}{abbreviate name or not (default: FALSE)}

\item{minlength}{set minmal length for names}

\item{abbr.method}{abbreviation method(default: both side)}

\item{text.scale}{vector of text size for upset (ylab,yaxis,xlab,group name,xaxis,insection)}
}
\value{
venn object and figures
}
\description{
Extracts shared and unique elements from groups provided to the function.
This base function will create a formal class venn object and can also graphically plot the amount
of objects in each group. The plot will be in the form of a tradition venn diagram with default. And user
can also use vennpie or upset method to display the result.
}
\details{
Extract shared and unique sets
}
\examples{
A<-sample(1:100, 40, replace = FALSE);
B<-sample(1:100, 60, replace = FALSE);
C<-sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
}
\author{
Kai Guo
}
