% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{vennpie}
\alias{vennpie}
\title{pie plot show shared and unique part}
\usage{
vennpie(object, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{set name you want display}

\item{color}{vector of color you want use for the group}

\item{revcolor}{backgroup color}

\item{show.number}{display the element number of the group or not}

\item{log}{use log transform or not}

\item{base}{log base}

\item{sep}{separate for new colnames}

\item{percentage}{display percentage format or not}
}
\value{
plot
}
\description{
Make vennpie figure
}
\author{
Kai Guo
}
