% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/vennpie.R
\docType{methods}
\name{vennpie}
\alias{vennpie}
\title{pie plot show shared and unique part}
\usage{
vennpie(object, group = NULL, color = NULL, revcolor = "lightgrey",
  any = NULL, show.number = TRUE, show.x = TRUE, sep = "_",
  log = FALSE, base = NULL, percentage = FALSE, ...)

\S4method{vennpie}{venn}(object, group = NULL, color = NULL,
  revcolor = "lightgrey", any = NULL, show.number = TRUE,
  show.x = TRUE, sep = "_", log = FALSE, base = NULL,
  percentage = FALSE, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{set name you want display}

\item{color}{vector of color you want use for the group}

\item{revcolor}{backgroup color}

\item{show.number}{display the element number of the group or not}

\item{sep}{separate for new colnames}

\item{log}{use log transform or not}

\item{base}{log base}

\item{percentage}{display percentage format or not}

\item{object}{venn object}

\item{group}{set name user want to highlight}

\item{color}{vector of color you want use for the group}

\item{revcolor}{color user want for other sets except highlight}

\item{show.number}{display the element number of the group or not}

\item{log}{use log transform or not}

\item{base}{log base}

\item{sep}{separate for new colnames}

\item{percentage}{display percentage format or not}
}
\value{
vennpie graphic
}
\description{
Vennpie uses the venn object and  create a figure in the form of a venn pie diagram
rather than a traditional venn diagram. User are allowed to highlight a specific section of the venn pie.
With revcolor paramater user can set the color of all groups not selected using the group argument.
}
\examples{
A<-sample(1:100, 40, replace = FALSE);
B<-sample(1:100, 60, replace = FALSE);
C<-sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
vennpie(res)
}
\author{
Kai Guo
}
