% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/vennpie.R
\docType{methods}
\name{vennpie}
\alias{vennpie}
\title{Pie plot show shared and unique set}
\usage{
vennpie(object, group = NULL, color = NULL, revcolor = "lightgrey",
  any = NULL, show.number = TRUE, show.x = TRUE, sep = "_",
  log = FALSE, base = NULL, percentage = FALSE, ...)

\S4method{vennpie}{venn}(object, group = NULL, color = NULL,
  revcolor = "lightgrey", any = NULL, show.number = TRUE,
  show.x = TRUE, sep = "_", log = FALSE, base = NULL,
  percentage = FALSE, ...)
}
\arguments{
\item{object}{venn object}

\item{group}{set names user want to highlight}

\item{color}{vector of color you want to use for the sets}

\item{revcolor}{color user want for the non-selected sets}

\item{any}{alternative ways to highlight sets (vennpie)}

\item{show.number}{display the element number of the group or not}

\item{show.x}{show label outside the circle (default: TRUE)}

\item{sep}{separation character for new colnames}

\item{log}{log transform (default: FALSE)}

\item{base}{base value for log transformation}

\item{percentage}{display percentage format (default: FALSE)}

\item{piecolor}{colors for vennpie}

\item{sets.x.label}{x-axis label (upset)}
}
\value{
vennpie graphic
}
\description{
Vennpie uses the venn object and to create a figure in the form of a venn pie diagram
rather than a traditional venn diagram. User are allowed to highlight a specific section of the venn pie.
Using the revcolor paramater user can set the color of all groups not selected using the group argument.
}
\examples{
A<-sample(1:100, 40, replace = FALSE);
B<-sample(1:100, 60, replace = FALSE);
C<-sample(1:100, 40, replace = FALSE);
res<-venndetail(list(A=A,B=B,C=C),plot=TRUE)
vennpie(res)
}
\author{
Kai Guo
}
